move recording render pass to contents.
