+++ impeller/aiks/testing/context_spy.cc
@@ -2,9 +2,9 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-#include "impeller/aiks/testing/context_spy.h"
 #include <memory>
-#include "impeller/aiks/testing/recording_render_pass.h"
+
+#include "impeller/aiks/testing/context_spy.h"
 
 namespace impeller {
 namespace testing {
