+++ impeller/entity/contents/test/recording_render_pass.cc
@@ -2,7 +2,7 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-#include "impeller/aiks/testing/recording_render_pass.h"
+#include "impeller/entity/contents/test/recording_render_pass.h"
 
 #include <utility>
